modules/location-event-microservice/requirements.txt
grpcio == 1.33.2
grpcio-tools == 1.33.2
kafka-python